feat: add routing and update dependencies
- Added react-router-dom for routing functionality.
- Updated @vitejs/plugin-react to version 6.0.1.
- Created a 404.html file for handling unknown routes on static hosts.
- Refactored App component to use HashRouter and Routes for navigation.
- Moved home page content to a new HomePage component.
- Added NotFoundPage component for handling 404 errors.
- Updated styles for the NotFoundPage.
- Fixed a typo in assets README.md.
