+++ src/main.jsx
@@ -3,7 +3,18 @@ import { createRoot } from "react-dom/client";
 import App from "./App";
 import "./styles/app.css";
 
-createRoot(document.getElementById("root")).render(
+if (window.location.pathname !== "/" && !window.location.hash) {
+  const route = window.location.pathname + window.location.search;
+  window.history.replaceState(null, "", "/#" + route);
+}
+
+const rootElement = document.getElementById("root");
+
+if (!rootElement) {
+  throw new Error("Root container not found");
+}
+
+createRoot(rootElement).render(
   <React.StrictMode>
     <App />
   </React.StrictMode>
